# perf: optimize bulk encode decode