--- a/Cargo.toml
+++ b/Cargo.toml
@@ -132,7 +132,7 @@ etcd-client = "0.14"
 fst = "0.4.7"
 futures = "0.3"
 futures-util = "0.3"
-greptime-proto = { git = "https://github.com/GreptimeTeam/greptime-proto.git", rev = "893773a6eea87fb0de5465c89cca1d1a0ba38efc" }
+greptime-proto = { git = "https://github.com/GreptimeTeam/greptime-proto.git", rev = "2a245c8962b73809cfbe66266d463864864c0141" }
 hex = "0.4"
 http = "1"
 humantime = "2.1"