--- a/src/operator/src/bulk_insert.rs
+++ b/src/operator/src/bulk_insert.rs
@@ -98,8 +98,8 @@ impl Inserter {
                     body: Some(bulk_insert_request::Body::ArrowIpc(ArrowIpc {
                         region_id: region_id.as_u64(),
                         schema: schema_bytes,
-                        payload: data.data_body,
                         data_header: data.data_header,
+                        payload: data.data_body,
                     })),
                 })),
             };
@@ -196,8 +196,8 @@ impl Inserter {
                                 body: Some(bulk_insert_request::Body::ArrowIpc(ArrowIpc {
                                     region_id: region_id.as_u64(),
                                     schema: schema_bytes,
-                                    payload,
                                     data_header: header,
+                                    payload,
                                 })),
                             })),
                         };