--- a/src/store-api/src/metadata.rs
+++ b/src/store-api/src/metadata.rs
@@ -1027,14 +1027,6 @@ pub enum MetadataError {
         #[snafu(implicit)]
         location: Location,
     },
-
-    #[snafu(display("Failed to decode prost message"))]
-    Prost {
-        #[snafu(source)]
-        error: prost::DecodeError,
-        #[snafu(implicit)]
-        location: Location,
-    },
 }
 
 impl ErrorExt for MetadataError {