--- a/src/store-api/src/region_request.rs
+++ b/src/store-api/src/region_request.rs
@@ -340,9 +340,9 @@ fn make_region_bulk_inserts(request: BulkInsertRequest) -> Result<Vec<(RegionId,
     let decoder_timer = metrics::CONVERT_REGION_BULK_REQUEST
         .with_label_values(&["decode"])
         .start_timer();
-    let mut decoder = FlightDecoder::try_from_schema_bytes(schema).unwrap();
+    let mut decoder = FlightDecoder::try_from_schema_bytes(&schema).unwrap();
     let payload = decoder
-        .try_decode_record_batch(data_header, payload)
+        .try_decode_record_batch(&data_header, &payload)
         .unwrap();
     decoder_timer.observe_duration();
     let region_id: RegionId = region_id.into();