I hereby agree to the terms of the [GreptimeDB CLA](https://github.com/GreptimeTeam/.github/blob/main/CLA.md).

## Refer to a related PR or issue link (optional)
- Please don't merge until https://github.com/GreptimeTeam/greptime-proto/pull/240 is merged.
## What's changed and what's your intention?

 - **Update Flight Decoder**: Enhanced `FlightDecoder` in `flight.rs` to support decoding from schema bytes and added methods for decoding record batches.
 - **Modify Bulk Insert Logic**: Updated `bulk_insert.rs` to utilize the new `FlightDecoder` methods for handling schema and payload data.
 - **Add Benchmarks**: Introduced a new benchmark for the flight decoder in `bench_flight_decoder.rs` and updated `bench_main.rs` to include it.

## Benchmark
### Endode performance
If the bulk insert request only involves one region, then we can spare one encode on frontend and postpone the decoding to datanode, which can save x10ms encoding time:

<img width="2497" alt="image" src="https://github.com/user-attachments/assets/c7af05e1-bdcd-4b41-8168-16e35f4fc3f5" />


### Decode performance
Slightly improve decode performance by ~5%

```
$ cargo bench -p common-grpc -- 'flight_decoder'
flight_decoder_100000_rows/decode_from_protobuf
                        time:   [1.6648 ms 1.6657 ms 1.6666 ms]
                        change: [+0.3603% +0.4654% +0.6247%] (p = 0.00 < 0.05)
                        Change within noise threshold.
flight_decoder_100000_rows/decode_from_header_and_body
                        time:   [1.6186 ms 1.6199 ms 1.6215 ms]
                        change: [-0.1881% -0.0506% +0.0712%] (p = 0.47 > 0.05)
                        No change in performance detected.
flight_decoder_200000_rows/decode_from_protobuf
                        time:   [3.3856 ms 3.3921 ms 3.3995 ms]
                        change: [-0.7051% -0.3268% +0.0194%] (p = 0.08 > 0.05)
                        No change in performance detected.
flight_decoder_200000_rows/decode_from_header_and_body
                        time:   [3.2001 ms 3.2149 ms 3.2325 ms]
                        change: [+0.3708% +0.8536% +1.5174%] (p = 0.00 < 0.05)
                        Change within noise threshold.
flight_decoder_1000000_rows/decode_from_protobuf
                        time:   [20.552 ms 20.581 ms 20.621 ms]
flight_decoder_1000000_rows/decode_from_header_and_body
                        time:   [18.686 ms 18.742 ms 18.802 ms]
```

This also reflects in the decode stage elapsed time:

<img width="1183" alt="image" src="https://github.com/user-attachments/assets/aca171b6-f1ba-4d15-8d5a-520a305cfc19" />


## PR Checklist
Please convert it to a draft if some of the following conditions are not met.

- [x] I have written the necessary rustdoc comments.
- [x] I have added the necessary unit tests and integration tests.
- [ ] This PR requires documentation updates.
- [x] API changes are backward compatible.
- [x] Schema or data changes are backward compatible.